Restructure `io` module
- minimize imports of other package (XLSX stuff only when needed).
- separate interdependent utilties from "leaf" helpers
- remove obsolete todos from documentation
